Removes getters in favor of public member variables.
Also makes `requestReturn` external (it is good to be explicit about these things).

Since this contract was put together in a hurry, it is best to keep it as simple as possible, with no extra features beyond what is necessary to do its job.  In this case, the getters were unnecessary complications that could be achieve (mostly) through a language feature.  This won't do the defaulting for you, but the defaulting is just as easy to do in the calling contract/script as it is here.