Adds return statement to function.
Having a return statement allows the caller to assert that the call was made and was successful.  Without a return statement, the caller can't tell the difference between the call running out of gas, the input address being wrong, the function signature being wrong (older versions of Solidity) and a successful call.  With a return statement the caller can do `require(chooseWhgReturnAddress.requestReturn(msg.sender));` to ensure safety.